[Refactor] #65 Response 하위 패키지 수정
