+++ auth-service/src/main/java/com/ticketrush/boundedcontext/auth/app/dto/response/login/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.ticketrush.boundedcontext.auth.app.dto.response.signup;
+package com.ticketrush.boundedcontext.auth.app.dto.response.login;
 
 public record LoginResponse(
     Long userId, String email, String role, String accessToken, String refreshToken) {}
