+++ auth-service/src/main/java/com/ticketrush/boundedcontext/auth/app/usecase/LoginUseCase.java
@@ -1,8 +1,8 @@
 package com.ticketrush.boundedcontext.auth.app.usecase;
 
 import com.ticketrush.boundedcontext.auth.app.dto.request.LoginRequest;
-import com.ticketrush.boundedcontext.auth.app.dto.response.signup.LoginResponse;
-import com.ticketrush.boundedcontext.auth.app.dto.response.signup.UserServiceAuthInfoResponse;
+import com.ticketrush.boundedcontext.auth.app.dto.response.login.LoginResponse;
+import com.ticketrush.boundedcontext.auth.app.dto.response.login.UserServiceAuthInfoResponse;
 import com.ticketrush.boundedcontext.auth.out.apiclient.UserServiceClient;
 import com.ticketrush.global.exception.BusinessException;
 import com.ticketrush.global.security.JwtTokenProvider;
